refactor: move major code inside lib folder (#33)
* refactor: move major code inside lib folder

* refacotr

* refactor
