+++ src/lib/api/appwrite/boards.api.ts
@@ -9,12 +9,12 @@ import { db, storage } from './client';
 import { v4 as uuidv4 } from 'uuid';
 import APPWRITE_CONST from '$constants/appwrite.constants';
 import { Query } from 'appwrite';
-import boardStore from '$lib/store/boards.store';
+import boardStore from '$store/boards.store';
 import toast from 'svelte-french-toast';
 import type { BoardDescriptionFormValues } from '$types/formValues';
 import TEXT from '$constants/text.constants';
 import { getBulkUserData } from './userDetails.api';
-import { enhanceBoardData } from '$lib/transformers/board.transformer';
+import { enhanceBoardData } from '$transformers/board.transformer';
 
 const uploadBoardCover = async (file: File): Promise<string> => {
 	const fileId = uuidv4();
