+++ src/lib/constants/appwrite.constants.ts
