+++ src/lib/constants/text.constants.ts
