+++ src/lib/enums/Input.enums.ts
