+++ src/lib/types/appwriteDocs.types.ts
