+++ src/lib/types/authStore.d.ts
