+++ src/lib/types/board.d.ts
