+++ src/lib/types/error.types.ts
