+++ src/lib/utils/getNavType.utils.ts
