+++ src/routes/(auth)/login/+page.ts
@@ -1,5 +1,5 @@
 import type { PageLoad } from './$types';
-import ROUTES from '$constants/routes.constants';
+import ROUTES from '$lib/constants/routes.constants';
 
 // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
 export const load = (({ url }) => {
