+++ svelte.config.js
@@ -24,15 +24,16 @@ const config = {
 			$lib: 'src/lib',
 			$components: 'src/lib/components',
 			$sass: 'src/lib/sass',
-			$types: 'src/types',
+			$types: 'src/lib/types',
 			$helpers: 'src/lib/helpers',
-			$enums: 'src/enums',
-			$utils: 'src/utils',
-			$constants: 'src/constants',
+			$enums: 'src/lib/enums',
+			$utils: 'src/lib/utils',
+			$constants: 'src/lib/constants',
 			$validations: 'src/lib/validations',
 			$store: 'src/lib/store',
 			$api: 'src/lib/api',
 			$factories: 'src/lib/factories',
+			$transformers: 'src/lib/transformers',
 		},
 	},
 };
